+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 
 ## [2.77.0] 2022-04-05
 
+- hardis:org:user:activateinvalid : new --profiles argument
 - Update MsTeams WebHooks ENV variables
   - MS_TEAMS_WEBHOOK_URL
   - CRITICAL_MS_TEAMS_WEBHOOK_URL
@@ -13,6 +14,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
   - WARNING_MS_TEAMS_WEBHOOK_URL
   - INFO_MS_TEAMS_WEBHOOK_URL
 - Allow to install packages during deployment check using INSTALL_PACKAGES_DURING_CHECK_DEPLOY=true env variable
+- Enhance prompt org labels
 
 ## [2.76.2] 2022-04-04
 
@@ -24,7 +26,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 
 ## [2.76.0] 2022-04-03
 
-- New command **sfdx**hardis:org:user:activateinvalid** to activate invalid emails in sandbox
+- New command **sfdx hardis:org:user:activateinvalid** to activate invalid emails in sandbox
 - Fix CI org authentication in case the default username is not the org that we want to configure
 - Bypass error with force:source:legacy:pull / push
 - hardis:work:save : Propose to manually commit files